# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User login and register",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://165.22.71.144/uears_web/public/api",
            "description": "Base URL"
        }
    ],
    "paths": {
...
        "/insert_faculty": {
            "post": {
                "operationId": "App\\Http\\Controllers\\API\\FacultyController::insert_faculty",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "faculty name",
                        "required": true
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "image of faculty",
                        "required": true
                    },
                    {
                        "name": "info",
                        "in": "query",
                        "description": "info of faculty",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "faculty description",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "insert faculty with above params"
                    }
                }
            }
        },
...
        "/get_faculty": {
            "get": {
                "operationId": "App\\Http\\Controllers\\API\\FacultyController::get_faculty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "faculty id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get one faculty with its id"
                    }
                }
            }
        },
...
        "/insert_venue": {
            "post": {
                "operationId": "App\\Http\\Controllers\\API\\VenueController::insert_venue",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "location latitude",
                        "required": true
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "location longitude",
                        "required": true
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "location as a string",
                        "required": true
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "mobile phone number of location",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "description of location",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "insert a new location with above inputs"
                    }
                }
            }
        },
        "/display_venue": {
            "get": {
                "operationId": "App\\Http\\Controllers\\API\\VenueController::display_venue",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "location latitude",
                        "required": true
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "location longitude",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "display a specific venue"
                    }
                }
            }
        }
    }
}